[Subtyping Generator] Sync `feature/subtype-gen` branch with `master`

+++ interpreter/value_accountcapabilitycontroller.go
@@ -152,7 +152,7 @@ func (*AccountCapabilityControllerValue) IsStorable() bool {
 func (v *AccountCapabilityControllerValue) Storable(
 	storage atree.SlabStorage,
 	address atree.Address,
-	maxInlineSize uint64,
+	maxInlineSize uint32,
 ) (
 	atree.Storable,
 	error,
